pyproject.toml
[tool.poetry]
name = "mtgcobuilderapi"
version = "0.1.0"
description = "API for MTG Cobuilder"
authors = ["Kamil Rybacki <kamilandrzejrybacki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = { version="^2.0.16", extras=["asyncio"]}
asyncpg = "^0.30.0"
environ-config = "^24.1.0"
nest-asyncio = "^1.6.0"
httpx = "^0.28.1"
tenacity = "^9.1.2"
dependency-injector = "^4.46.0"

[tool.poetry.group.dev.dependencies]
mypy = "1.15.0"
ruff = "0.11.8"
pre-commit = "^3.6.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
pytest-cov = "^4.1.0"
pytest-order = "^1.3.0"
testcontainers = "^4.10.0"
types-requests = "^2.32.4.20250611"
imagehash = "^4.3.2"
pillow = "^11.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN",  # flake8-annotations
    "COM",  # flake8-commas
    "C90",  # mccabe complexity
    "DJ",   # django
    "EXE",  # flake8-executable
    "T10",  # debugger
    "TID",  # flake8-tidy-imports
    "D100",   # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D205",
    "D203",
    "D212",
    "D400",
    "D401",
    "D415",
    "E402",   # false positives for local imports
    "E501",   # line too long
    "TRY003", # external messages in exceptions are too verbose
    "TD002",
    "TD003",
    "FIX002", # too verbose descriptions of todos
    "PGH003", # concrete type ignores
    "UP045",  # for BaseModel use Optional[X] instead of X | None
    "G004",
    "EM101",
    "EM102",
    "LOG015",
    "S101",
    "ASYNC210",
    "TC001",
    "SIM102",
    "PLR0911",
    "SLF001",
    "FBT001",
    "FBT002",
    "UP046",
    "D413",
    "D412",
]
exclude = [
    "*.pyi",  # ignore type stubs
    "*.ipynb",  # ignore Jupyter notebooks
    "tests/**",  # ignore tests
    "make/**",  # ignore types compiler and other build utils
]

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
warn_unused_ignores = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
show_error_codes = true
show_column_numbers = true
pretty = true

[tool.pytest.ini_options]

asyncio_default_fixture_loop_scope = "session"
markers = [
    "offline: tests that can be carried out offline",
    "chosen: helper marker for running only chosen tests with `test-chosen` command",
]
